auto-complete/requirements.txt
Package                       Version
----------------------------- ---------
absl-py                       0.10.0
aiohttp                       3.8.1
aiosignal                     1.2.0
anyio                         3.4.0
appnope                       0.1.2
argon2-cffi                   21.3.0
argon2-cffi-bindings          21.2.0
astunparse                    1.6.3
async-generator               1.10
async-timeout                 4.0.2
attrs                         21.4.0
Babel                         2.9.1
backcall                      0.2.0
backports.functools-lru-cache 1.6.4
bleach                        4.1.0
blinker                       1.4
brotlipy                      0.7.0
cached-property               1.5.2
cachetools                    4.2.4
certifi                       2021.10.8
cffi                          1.15.0
charset-normalizer            2.0.9
click                         8.0.3
cryptography                  36.0.1
cycler                        0.11.0
dataclasses                   0.8
debugpy                       1.5.1
decorator                     5.1.0
defusedxml                    0.7.1
entrypoints                   0.3
flatbuffers                   2.0
flit_core                     3.6.0
fonttools                     4.28.5
frozenlist                    1.2.0
gast                          0.4.0
google-auth                   1.35.0
google-auth-oauthlib          0.4.6
google-pasta                  0.2.0
grpcio                        1.43.0
h5py                          3.1.0
idna                          3.1
importlib-metadata            4.10.0
importlib-resources           5.4.0
ipykernel                     6.6.1
ipython                       7.31.0
ipython-genutils              0.2.0
jedi                          0.18.1
Jinja2                        3.0.3
joblib                        1.1.0
json5                         0.9.5
jsonschema                    4.3.3
jupyter-client                7.1.0
jupyter-core                  4.9.1
jupyter-server                1.13.1
jupyterlab                    3.2.5
jupyterlab-pygments           0.1.2
jupyterlab-server             2.10.2
keras                         2.7.0
Keras-Preprocessing           1.1.2
kiwisolver                    1.3.2
libclang                      12.0.0
Markdown                      3.3.6
MarkupSafe                    2.0.1
matplotlib                    3.5.1
matplotlib-inline             0.1.3
mistune                       0.8.4
multidict                     5.2.0
munkres                       1.1.4
nbclassic                     0.3.4
nbclient                      0.5.9
nbconvert                     6.4.0
nbformat                      5.1.3
nest-asyncio                  1.5.4
notebook                      6.4.6
numpy                         1.19.5
oauthlib                      3.1.1
olefile                       0.46
opt-einsum                    3.3.0
packaging                     21.3
pandas                        1.3.5
pandocfilters                 1.5.0
parso                         0.8.3
patsy                         0.5.2
pexpect                       4.8.0
pickleshare                   0.7.5
Pillow                        8.4.0
pip                           21.3.1
prometheus-client             0.12.0
prompt-toolkit                3.0.24
protobuf                      3.19.1
ptyprocess                    0.7.0
pyasn1                        0.4.8
pyasn1-modules                0.2.7
pycparser                     2.21
Pygments                      2.11.1
PyJWT                         2.3.0
pyOpenSSL                     21.0.0
pyparsing                     3.0.6
pyrsistent                    0.18.0
PySocks                       1.7.1
python-dateutil               2.8.2
pytz                          2021.3
pyu2f                         0.1.5
pyzmq                         22.3.0
requests                      2.27.0
requests-oauthlib             1.3.0
rsa                           4.8
scikit-learn                  1.0.2
scipy                         1.7.3
seaborn                       0.11.2
Send2Trash                    1.8.0
setuptools                    60.2.0
six                           1.15.0
sniffio                       1.2.0
statsmodels                   0.13.1
tensorboard                   2.6.0
tensorboard-data-server       0.6.0
tensorboard-plugin-wit        1.8.1
tensorflow-estimator          2.7.0
tensorflow-macos              2.7.0
termcolor                     1.1.0
terminado                     0.12.1
testpath                      0.5.0
threadpoolctl                 3.0.0
tornado                       6.1
traitlets                     5.1.1
typing-extensions             3.7.4.3
unicodedata2                  14.0.0
urllib3                       1.26.7
wcwidth                       0.2.5
webencodings                  0.5.1
websocket-client              1.2.3
Werkzeug                      2.0.1
wheel                         0.35.1
wrapt                         1.12.1
yarl                          1.7.2
zipp                          3.6.0
